requirements.txt
python-binance>=1.0.35
pandas>=2.0
pandas-ta>=0.4.0b
numpy>=1.24
rich>=13.0
aiohttp>=3.9
websockets>=12.0
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.21
